refactor: Update Home page to use ProductCard component and fetch product data
This commit refactors the Home page to use the newly added ProductCard component for displaying product information. It also fetches product data using the fetcher functions added in a previous commit. This improves the modularity and reusability of the code, as well as enhances the user experience by dynamically rendering product cards on the client side.